Correct type casting and formatting in Interrupt.c
Fix formatting and add type casting for variable 'i'.

+++ src/AmtPtpDeviceUsbKm/Interrupt.c
@@ -274,7 +274,7 @@ AmtPtpEvtUsbInterruptPipeReadComplete(
 						TraceEvents(
 							TRACE_LEVEL_INFORMATION, TRACE_INPUT,
 							"%!FUNC! Contact %d lifted",
-							i
+							(ULONG)i
 						);
 						reportSlots++;
 					}
